// .tmp/backups/code_snapshot/types.ts
export interface Curso {
  id: string;
  tenantId: string;
  nome: string;
  cargaHoraria?: string; // Ex: "40h"
  cor: string;
  minutosPorHora?: number; // 50 ou 60
  numeroCurso?: string; // Identificador Externo
  status: 'ativo' | 'concluido';
}